introduce JointThreadAllocInfo
