+++ dbms/src/Storages/KVStore/KVStore.h
@@ -18,6 +18,7 @@
 #include <Storages/DeltaMerge/DeltaMergeInterfaces.h>
 #include <Storages/DeltaMerge/RowKeyRange.h>
 #include <Storages/KVStore/Decode/RegionDataRead.h>
+#include <Storages/KVStore/FFI/JointThreadAllocInfo.h>
 #include <Storages/KVStore/MultiRaft/Disagg/RaftLogManager.h>
 #include <Storages/KVStore/MultiRaft/PreHandlingTrace.h>
 #include <Storages/KVStore/MultiRaft/RegionManager.h>
@@ -47,9 +48,6 @@ namespace tests
 {
 class KVStoreTestBase;
 } // namespace tests
-
-enum class ReportThreadAllocateInfoType : uint64_t;
-struct ReportThreadAllocateInfoBatch;
 
 class IAST;
 using ASTPtr = std::shared_ptr<IAST>;
@@ -122,26 +120,6 @@ struct ProxyConfigSummary
     size_t snap_handle_pool_size = 0;
 };
 
-struct ThreadInfoJealloc
-{
-    uint64_t allocated_ptr{0};
-    uint64_t deallocated_ptr{0};
-
-    uint64_t allocated() const
-    {
-        if (allocated_ptr == 0)
-            return 0;
-        return *reinterpret_cast<uint64_t *>(allocated_ptr);
-    }
-    uint64_t deallocated() const
-    {
-        if (deallocated_ptr == 0)
-            return 0;
-        return *reinterpret_cast<uint64_t *>(deallocated_ptr);
-    }
-    int64_t remaining() const { return static_cast<int64_t>(allocated()) - static_cast<int64_t>(deallocated()); }
-};
-
 /// KVStore manages raft replication and transactions.
 /// - Holds all regions in this TiFlash store.
 /// - Manages region -> table mapping.
@@ -171,9 +149,7 @@ public:
     // Proxy will validate and refit the config items from the toml file.
     const ProxyConfigSummary & getProxyConfigSummay() const { return proxy_config_summary; }
     void reportThreadAllocInfo(std::string_view, ReportThreadAllocateInfoType type, uint64_t value);
-    static void reportThreadAllocBatch(std::string_view, ReportThreadAllocateInfoBatch data);
-    void recordThreadAllocInfo();
-    void stopThreadAllocInfo();
+    void reportThreadAllocBatch(std::string_view, ReportThreadAllocateInfoBatch data);
 
 public: // Region Management
     void restore(PathPool & path_pool, const TiFlashRaftProxyHelper *);
@@ -438,13 +414,7 @@ private:
     std::atomic<int64_t> ongoing_prehandle_task_count{0};
     ProxyConfigSummary proxy_config_summary;
 
-    mutable std::shared_mutex memory_allocation_mut;
-    std::unordered_map<std::string, ThreadInfoJealloc> memory_allocation_map;
-
-    bool is_terminated{false};
-    mutable std::mutex monitoring_mut;
-    std::condition_variable monitoring_cv;
-    std::thread * monitoring_thread{nullptr};
+    JointThreadInfoJeallocMapPtr joint_memory_allocation_map;
 };
 
 /// Encapsulation of lock guard of task mutex in KVStore
